base_page added shop_user_model modified login_creds.json removed Readme updated
