+++ pages/std_user_inventory_page.py
@@ -1,5 +1,6 @@
 from selenium.webdriver.common.by import By
-from pages.login_page import BasePage
+from pages.base_page import BasePage
+
 
 BACKPACK_ADD_TO_CART_LOCATOR = (By.ID, 'add-to-cart-sauce-labs-backpack')
 TSHIRT_ADD_TO_CART_LOCATOR = (By.ID, 'add-to-cart-sauce-labs-bolt-t-shirt')
@@ -21,7 +22,7 @@ class StdUserInvPage(BasePage):
 
     def go_to_cart(self):
         self.driver.find_element(By. CLASS_NAME, 'shopping_cart_link').click()
-        return StdUserCartPage(self.driver)
+        # return StdUserCartPage(self.driver)
