+++ technical/shop_user_model.py
@@ -2,15 +2,14 @@ from dataclasses import dataclass
 
 
 @dataclass
-class ShopUser:
-    name: str
-    surname: str
-    zip_code: str
+class UserName:
+    user_1: str = "standard_user"
+    user_2: str = "locked_out_user"
+    user_3: str = "problem_user"
+    user_4: str = "performance_glitch_user"
 
 
+@dataclass
+class Password:
+    pwd: str = "secret_sauce"
 
-x = {"name": "Marcin", "surname": "Kowal"}
-print(x["namr"])
-
-user = ShopUser("Marcin", "Kowal", "30-123")
-print(user.name)
