+++ tests/test_buy_2_products.py
